+++ src/core/engine.h
@@ -14,7 +14,7 @@
 typedef struct EngineTimer {
     struct timeval started_at;
     struct timeval paused_at;
-    uint32_t pause_skip;
+    unsigned long pause_skip;
     int paused;
 } EngineTimer;
 
@@ -44,7 +44,7 @@ void Engine_regulate(Engine *engine);
 
 void Engine_pause_time(Engine *engine);
 void Engine_resume_time(Engine *engine);
-uint32_t Engine_get_ticks(Engine *engine);
+unsigned long Engine_get_ticks(Engine *engine);
 
 struct Scene;
 struct Scene *Engine_get_current_scene(Engine *engine);
