+++ src/core/gzip.c
@@ -5,14 +5,14 @@
 #include <lcthw/bstrlib.h>
 #include "../dbg.h"
 
-unsigned long int decompress_data(unsigned char* abSrc, int nLenSrc, unsigned char **abDst,
+unsigned long int decompress_data(unsigned char* abSrc, unsigned long nLenSrc, unsigned char **abDst,
                     int nLenDst) {
   z_stream zInfo = {.avail_in = 0};
   int prevSize = nLenDst;
   unsigned char *src_start = abSrc;
   unsigned char *dst_start = *abDst;
 
-  zInfo.total_in = zInfo.avail_in =  nLenSrc;
+  zInfo.total_in = zInfo.avail_in = (int)nLenSrc;
   zInfo.total_out = zInfo.avail_out = nLenDst;
   zInfo.next_in = src_start;
   zInfo.next_out = dst_start;
